Fix interactive mode escape sequence
There were two problems:

1. Escape sequence is a regular expression, therefore '.' must be prefixed
   with '\'.
2. Delete key produces an escape sequence that ends with '~'. Therefor
   pressing Delete followed by '.' exits novaboot.

It would be nice if we can use traditional esscape sequence '~.' at the
beginning of the line (as in ssh or cu), but it seems that Expect.pm cannot
handle the "beginning of the line" in the escape sequence correctly.
Therefore we use "~~." sequence to solve problem 2.