U-Boot: Extend kernel command line handling
In order to be compatible with a standard U-Boot way of having
variables containin U-Boot commands and using 'run' to expand variable
references in them, we check kernel command line if it starts with
'setenv bootargs'. If yes, we simply run this command, otherwise, we
preprend 'setenv bootargs' ourselves.

Examples:

1) Novaboot script contains:

       load linux console=${console} quiet

   U-Boot is instructed to run:

       setenv bootargs console=${console} quiet

2) Novaboot script contains:

       load linux ${netargs} quiet

   and the U-Boot variable netargs contsins "setenv bootargs console=${console}"
   U-Boot is instructed to run:

       setenv bootargs console=${console} quiet

Note that in case 1) novaboot prepends "setenv bootargs" itselv
whereas in 2) inly executes the command line as is.