Warn when multiple target connection options are given
This may be problem when --remote-expect is used, but the expected string
is not produced by all connection methods. Then, --remote-expect will may
cause indefinite blocking. For example, when

   --remote-cmd 'sterm ...' --remote-exepect Connected

is overridedn with --serial ..., then the "Connected" may not be
received.

Since this override may behave correctly in certain cases, we only warn
here.

Thanks to Michal Vokáč for reporting this.