Add systemd units to power-on/off the target
The idea is that whenever somebody logs as a certain user, systemd switches
the target on. After all sessions of that user are closed, systemd starts
a timer and after its expiration, the target is switched off.

Not that power-on is not always working correctly. The reason is that
we would want the novaboot-target-on@XXX.service to finish before the user
is logged. This can be achieved by adding Before= directive to it, but the
problem is that user units do have UIDs and not names, so we cannot simply
write Before=user@%i.service.

Currently, I'm not sure how to properly solve this issue so the power-on
command is commented out. Power-off, however, work correctly.