server: Make console the default command
This is useful, because SSH does not properly terminate the console
command if it is not run on terminal. The command is run on terminal
when ssh is run with -t argument or when there is no command
specified. Hence, with this change, we it will be less likely that
users connect to the console and forget to specify -t.