server: Refactor delayed power-off logic
From time to time, I'm facing problems with delayed power off not
working. I'm not sure what is the exact reason, but a potential reason
could be that the board was powered on but the systemd service
novaboot-delayed-power-off.service, which is responsible for powering
off, was not started. To prevent that, we change the logic so that
powering on happens as a start job of the same service.